Inference queues
