+++ crates/cairo-lang-semantic/src/expr/test_data/snapshot
@@ -45,7 +45,7 @@ error: Unexpected argument type. Expected: "core::felt252", found: "@?1".
     let y: felt252 = @6;
                      ^^
 
-error: Trait has no implementation in context: core::integer::NumericLiteral::<@core::felt252>
+error: Type mismatch: core::felt252 and core::felt252
  --> lib.cairo:2:23
     let x: @felt252 = 5;
                       ^
