+++ crates/cairo-lang-semantic/src/items/imp.rs
@@ -36,6 +36,7 @@ use crate::db::SemanticGroup;
 use crate::diagnostic::SemanticDiagnosticKind::{self, *};
 use crate::diagnostic::{NotFoundItemType, SemanticDiagnostics};
 use crate::expr::compute::{compute_root_expr, ComputationContext, Environment};
+use crate::expr::inference::conform::InferenceConform;
 use crate::expr::inference::{ImplVar, Inference, InferenceData, InferenceResult};
 use crate::items::function_with_body::get_implicit_precedence;
 use crate::items::functions::ImplicitPrecedence;
