+++ crates/cairo-lang-semantic/src/resolve/mod.rs
@@ -24,6 +24,7 @@ use crate::corelib::core_module;
 use crate::db::SemanticGroup;
 use crate::diagnostic::SemanticDiagnosticKind::*;
 use crate::diagnostic::{NotFoundItemType, SemanticDiagnostics};
+use crate::expr::inference::conform::InferenceConform;
 use crate::expr::inference::{Inference, InferenceData};
 use crate::items::enm::SemanticEnumEx;
 use crate::items::functions::{GenericFunctionId, ImplGenericFunctionId};
